requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
ruff==0.0.263
Pillow==10.0.0
pytest==7.2.2
pytest-django==4.5.2